+++ SOURCE/Cpp_Files/[Huy]_Command.cpp
@@ -16,18 +16,20 @@ void Command_5(Task task)
         GenerateNearlySortedData(&arr1[0], size);
     else
         GenerateReverseData(&arr1[0], size);
-    vector<int> arr2 = vector<int> (arr1);
+    vector<int> arr2 = vector<int>(arr1);
 
     exportArrayToFile(arr1, "input.txt");
 
-    //Print some initial information to console screen
+    // Print some initial information to console screen
     printCmd(task);
 
-    //comparison
-    int comparisons_1 = 0; double time_1 = 0.0;
+    // comparison
+    long long comparisons_1 = 0;
+    double time_1 = 0.0;
     Algo_Measuring[task.indexAlgo1](arr1, comparisons_1, time_1);
-    int comparisons_2 = 0; double time_2 = 0.0;
+    long long comparisons_2 = 0;
+    double time_2 = 0.0;
     Algo_Measuring[task.indexAlgo1](arr2, comparisons_2, time_2);
-    
+
     print2Terminal_c(comparisons_1, comparisons_2, time_1, time_2);
 }
