+++ SOURCE/Cpp_Files/[Minh]_Command.cpp
@@ -14,16 +14,17 @@ void Command_2(Task task)
         GenerateReverseData(array, n);
 
     vector<int> arr = arr2Vec(array, n); //? Convert to vector
-    exportArrayToFile(arr, "input.txt");        //? Write down the input to input.txt
+    exportArrayToFile(arr, "input.txt"); //? Write down the input to input.txt
     delete[] array;
 
-    //Print some initial information to console screen
+    // Print some initial information to console screen
     printCmd(task);
 
-    //Measure time - comparisons
-    int comparisons = 0; double time = 0;
+    // Measure time - comparisons
+    long long comparisons = 0;
+    double time = 0;
     Algo_Measuring[task.indexAlgo1](arr, comparisons, time);
 
-    //Print measured data(s) to console screen
+    // Print measured data(s) to console screen
     print2Terminal_a(comparisons, time, task.outPara);
 }
